refactor(histogram): rename ErrNoClosedIssues to ErrNoHistogramData
