+++ internal/service/model.go
@@ -3,6 +3,8 @@ package service
 import (
 	"errors"
 	"time"
+
+	"github.com/Go-Yadro-Group-1/Jira-Analyzer/internal/repository"
 )
 
 type ChartType string
@@ -65,4 +67,28 @@ type PriorityBar struct {
 
 type PriorityChart struct {
 	Bars []PriorityBar `json:"bars"`
+}
+
+const (
+	statsSecondsPerHour = 3600.0
+	statsDaysInWeek     = 7.0
+)
+
+func toProjectStats(raw repository.ProjectStats) ProjectStats {
+	var avgCompletionTimeHours float64
+	if raw.CountClosed > 0 {
+		avgCompletionTimeHours = float64(raw.TotalDurationClosed) /
+			(float64(raw.CountClosed) * statsSecondsPerHour)
+	}
+
+	return ProjectStats{
+		CountTotal:               raw.CountTotal,
+		CountOpen:                raw.CountOpen,
+		CountClosed:              raw.CountClosed,
+		CountReopened:            raw.CountReopened,
+		CountResolved:            raw.CountResolved,
+		CountInProgress:          raw.CountInProgress,
+		AvgCompletionTimeHours:   avgCompletionTimeHours,
+		AvgCreatedPerDayLastWeek: float64(raw.CountCreatedLastWeek) / statsDaysInWeek,
+	}
 }
